Add Jib Maven Plugin for Docker Image Build
This commit introduces the following changes to the codebase:

- Add the `jib-maven-plugin` from `com.google.cloud.tools` with version `3.4.1`.
- Configure the plugin to use the `openjdk:17-jdk-alpine` base image.
- Set the target Docker image to be pushed to `registry.hub.docker.com/aamirxshaikh/${project.artifactId}`.

Purpose:

The purpose of this commit is to integrate the Jib Maven Plugin into the project's build process. Jib is a tool that simplifies the process of building Docker images for Java applications without the need for a Dockerfile. By configuring the Jib Maven Plugin, the project can now build a Docker image directly from the compiled Java artifacts during the Maven build process.