Add Docker Configuration for API Gateway
This commit introduces the following changes to the codebase:

- Adds a new `application-docker.properties` file in the `src/main/resources` directory.
- Configures the server port to `8080` using the `server.port` property.
- Sets the Eureka Server URL to `http://root:root@discovery-server:8761/eureka` using the `eureka.client.serviceUrl.defaultZone` property.
- Configures the OAuth2 Resource Server JWT issuer URI to `http://keycloak:8080/realms/spring-boot-shopping-realm` using the `spring.security.oauth2.resourceserver.jwt.issuer-uri` property.
- Sets the Zipkin tracing endpoint to `http://zipkin:9411` using the `management.zipkin.tracing.endpoint` property.
- Specifies the Eureka Server service name as `discovery-server` using the `app.eureka-server` property.

Purpose:
The purpose of this commit is to introduce Docker-specific configuration properties for the API Gateway service. These properties are intended to be used when running the application in a Docker container environment. The configurations include setting up the Eureka Server URL, OAuth2 Resource Server JWT issuer URI, Zipkin tracing endpoint, and other relevant settings specific to the Docker environment. This commit ensures that the API Gateway service can properly connect to other services and components when deployed in a Docker-based infrastructure.